finishing

+++ core/app/Http/Controllers/UnitController.php
@@ -1,6 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Models\Satuan_Obat;
 use App\Models\Unit;
 use Illuminate\Http\Request;
 
@@ -38,8 +40,12 @@ class UnitController extends Controller
     public function update(Request $request,$id)
     {
         $unit = Unit::findOrFail($id);
+        $cek = Satuan_Obat::where('unit_id',$unit->id)->count();
+        if($cek  > 0){
+            return redirect()->back()->with('gagal','Data tidak bisa di ubah karena sudah digunakan !');
+        }
         $unit->nama = $request->nama;
-        $unit->jumlah_satuan_terkecil = $request->jumlah;
+        $unit->tingkat_satuan = $request->jumlah;
         $unit->save();
         return redirect(route('unit'))->with('success','Data Berhasil di ubah !');
 
@@ -48,7 +54,15 @@ class UnitController extends Controller
     public function destroy(Request $request)
     {
         $unit = Unit::findOrFail($request->id);
-        $unit->delete();
+
+        $cek = Satuan_Obat::where('unit_id',$unit->id)->count();
+
+        if($cek == 0){
+            $unit->delete();
+        }
+
+        return response()->json($cek);
+        // $unit->delete();
     }
 }
